Factored out most of a CKernelWriterClass. Still have a couple things to decouple though.
